+++ R/preprocess-codeanalysis.R
@@ -98,6 +98,7 @@ analyze_delayed_assign <- function(expr, results, locals, allowed_globals) {
 }
 
 analyze_loadd <- function(expr, results) {
+  expr <- ignore_ignore(expr)
   expr <- match.call(drake::loadd, as.call(expr))
   expr <- expr[-1]
   ht_set(results$loadd, analyze_strings(expr["list"]))
@@ -110,12 +111,14 @@ analyze_loadd <- function(expr, results) {
 }
 
 analyze_readd <- function(expr, results, allowed_globals) {
+  expr <- ignore_ignore(expr)
   expr <- match.call(drake::readd, as.call(expr))
   ht_set(results$readd, analyze_strings(expr["target"]))
   ht_set(results$readd, safe_all_vars(expr["target"]))
 }
 
 analyze_file_in <- function(expr, results) {
+  expr <- ignore_ignore(expr)
   x <- analyze_strings(expr[-1])
   x <- file.path(x)
   x <- encode_path(x)
@@ -123,6 +126,7 @@ analyze_file_in <- function(expr, results) {
 }
 
 analyze_file_out <- function(expr, results) {
+  expr <- ignore_ignore(expr)
   x <- analyze_strings(expr[-1])
   x <- file.path(x)
   x <- encode_path(x)
@@ -130,6 +134,7 @@ analyze_file_out <- function(expr, results) {
 }
 
 analyze_knitr_in <- function(expr, results) {
+  expr <- ignore_ignore(expr)
   files <- analyze_strings(expr[-1])
   lapply(files, analyze_knitr_file, results = results)
   ht_set(results$knitr_in, encode_path(files))